+++ src/io/ipc/write/common.rs
@@ -371,7 +371,7 @@ pub struct EncodedData {
 /// Calculate an 8-byte boundary and return the number of bytes needed to pad to 8 bytes
 #[inline]
 pub(crate) fn pad_to_64(len: usize) -> usize {
-    (((len + 63) & !63) - len) as usize
+    ((len + 63) & !63) - len
 }
 
 /// An array [`Chunk`] with optional accompanying IPC fields.
